+++ app/controllers/org_controller.rb
@@ -1,0 +1,31 @@
+class OrgController < ApplicationController
+
+
+
+
+
+
+
+
+
+
+
+
+    private
+
+    def org_params
+        params
+        .require(:org) # filters
+        .permit(:nome,
+                :email,
+                :instituto,
+                :curso,
+                :info,
+                :foto,
+                :link_site,
+                :link_fb,
+                :link_tt)
+    end
+
+
+end