+++ app/models/org.rb
@@ -1,0 +1,9 @@
+class Org < ApplicationRecord
+
+    validates :nome, presence: true, length: { in: 3..100 }
+    validates :email, presence: true, length: { maximum: 200 }, uniqueness: true
+    validates :instituto, presence: true
+    validates :curso, presence: true
+    validates :info, presence: true, length: { maximum: 1000 }
+
+end