+++ db/migrate/20190509183100_create_orgs.rb
@@ -1,0 +1,8 @@
+class CreateOrgs < ActiveRecord::Migration[5.2]
+  def change
+    create_table :orgs do |t|
+
+      t.timestamps
+    end
+  end
+end